+++ changelog.md
@@ -29,6 +29,7 @@ Codion Change Log
 - EntityTablePanel.configurePopupMenu() and configureToolBar() added.
 - EntityTablePanel bug fixed, Config.popupMenuEditAttributeControl() added.
 - EntityTablePanel bug fixed, now excludes the edit attribute control from the toolbar if an editPanel is available.
+- EntityEditPanel, EntityTablePanel.setupControls() renamed configureControls().
 
 ## 0.17.42
 ### is.codion.common.core